chore: rename docker-compose file
Prior to this change, the docker compose file was named
`compose-dev.yaml` meaning the -f flag and filename had to be provided
with every docker-compose command.

This change renames `compose-dev.yaml` to the conventional
`docker-compose.yaml` meaning the -f flaf and filename no longer have to
be provided when running docker-compose commands. This change speeds up
development.